fix: renamed file controller/fetchOrgas.ts to controller/getOrgas.ts for consistency
